\name{mlsas-package}
\alias{mlsas-package}\docType{package}
\title{Machine Learning Spatial Association Statistics}
\description{Implementation of learned attribute similarity and spatial randomization inference.}
\details{Start with \code{\link{mlsas_example}} and \code{\link{mlsas}}.
Default settings follow the empirical analysis in the accompanying paper.
The tutorial provides complete small-data workflows.}
\author{Woohyung Kim and Sang-Il Lee. Maintainer: Woohyung Kim \email{whk3423@snu.ac.kr}.}
